Extract getId / setId method from the MessageInterface to a new one
