+++ src/Message/ParametrizedMessageInterface.php
@@ -1,0 +1,17 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\Yii\Queue\Message;
+
+interface ParametrizedMessageInterface extends MessageInterface
+{
+    public function setId(?string $id): void;
+
+    /**
+     * Returns unique message ID.
+     *
+     * @return string|null
+     */
+    public function getId(): ?string;
+}
