+++ src/Middleware/FailureHandling/Implementation/SendAgainMiddleware.php
@@ -7,6 +7,7 @@ namespace Yiisoft\Yii\Queue\Middleware\FailureHandling\Implementation;
 use InvalidArgumentException;
 use Yiisoft\Yii\Queue\Message\Message;
 use Yiisoft\Yii\Queue\Message\MessageInterface;
+use Yiisoft\Yii\Queue\Message\ParametrizedMessageInterface;
 use Yiisoft\Yii\Queue\Middleware\FailureHandling\FailureHandlingRequest;
 use Yiisoft\Yii\Queue\Middleware\FailureHandling\MessageFailureHandlerInterface;
 use Yiisoft\Yii\Queue\Middleware\FailureHandling\MiddlewareFailureInterface;
@@ -40,11 +41,12 @@ final class SendAgainMiddleware implements MiddlewareFailureInterface
     ): FailureHandlingRequest {
         $message = $request->getMessage();
         if ($this->suites($message)) {
+            $id = $message instanceof ParametrizedMessageInterface ? $message->getId() : null;
             $message = new Message(
                 handlerName: $message->getHandlerName(),
                 data: $message->getData(),
                 metadata: $this->createMeta($message),
-                id: $message->getId(),
+                id: $id,
             );
             $message = $this->queue?->push($message) ?? $request->getQueue()->push($message);
